--- about/index.html
@@ -1,55 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>SOS2425-22 Team</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 40px;
-            padding: 20px;
-            background-color: #f4f4f4;
-            text-align: center;
-        }
-        .container {
-            background: white;
-            padding: 20px;
-            border-radius: 10px;
-            box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-            max-width: 600px;
-            margin: auto;
-        }
-        h1, h2, p {
-            margin: 5px 0;
-        }
-        ul {
-            list-style-type: none;
-            padding: 0;
-            margin: 5px 0;
-        }
-        li a {
-            text-decoration: none;
-            color: #007BFF;
-        }
-        li a:hover {
-            text-decoration: underline;
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <h1>SOS2425-22</h1>
-        <h2>Team</h2>
-        <ul>
-            <li><a href="https://github.com/juariolop1" target="_blank">Juan Manuel Ríos López</a></li>
-            <li><a href="https://github.com/gabiclone" target="_blank">Gabriel Cascajosa Rina</a></li>
-            <li><a href="https://github.com/aaleejandro01" target="_blank">Alejandro Morilla García</a></li>
-        </ul>
-        <p><strong>Project:</strong> Statistics of various crimes and the representative places of the Justice Ministry.</p>
-        <p><strong>Repository:</strong> <a href="https://github.com/gti-sos/SOS2425-22" target="_blank">gti-sos/SOS2425-22</a></p>
-        <p><strong>URL:</strong> <a href="https://sos2425-22.onrender.com/" target="_blank">https://sos2425-22.onrender.com/</a></p>
-        <p><strong>APIs:</strong> <a href="https://documenter.getpostman.com/view/42360434/2sB2cUC3wh" target="_blank">API JMRL</a>, <a href="https://documenter.getpostman.com/view/33043526/2sB2cUBP8d" target="_blank">API AMG</a>, TBD</p>
-    </div>
-</body>
-</html>
