--- public/index.html
@@ -1,98 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-  <title>SOS2425 - Presentación de APIs</title>
-  <style>
-    body {
-      font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-      margin: 0;
-      padding: 2rem;
-      background-color: #f0f2f5;
-      color: #333;
-    }
-
-    h1 {
-      text-align: center;
-      margin-bottom: 2rem;
-      color: #2c3e50;
-    }
-
-    .apis-container {
-      display: flex;
-      justify-content: center;
-      gap: 2rem;
-      flex-wrap: wrap;
-    }
-
-    .api-card {
-      background-color: #fff;
-      border-radius: 15px;
-      box-shadow: 0 4px 8px rgba(0,0,0,0.1);
-      width: 300px;
-      padding: 2rem 1rem 1rem;
-      text-align: center;
-      position: relative;
-    }
-
-    .api-title {
-      position: absolute;
-      top: -1.2rem;
-      left: 50%;
-      transform: translateX(-50%);
-      background-color: #3498db;
-      color: white;
-      padding: 0.4rem 1rem;
-      border-radius: 20px;
-      font-weight: bold;
-      box-shadow: 0 2px 5px rgba(0,0,0,0.1);
-    }
-
-    .description {
-      margin: 1.5rem 0;
-      font-size: 0.95rem;
-    }
-
-    a {
-      text-decoration: none;
-      color: #3498db;
-      font-weight: bold;
-    }
-
-    a:hover {
-      text-decoration: underline;
-    }
-
-    .footer {
-      text-align: center;
-      margin-top: 3rem;
-    }
-  </style>
-</head>
-<body>
-  <h1>Bienvenido a SOS2425</h1>
-
-  <div class="apis-container">
-    <div class="api-card">
-      <div class="api-title">Missing People Stats</div>
-      <p class="description">
-        Estadísticas sobre personas desaparecidas en España, clasificadas por provincia y año.
-      </p>
-      <a href="/api/v1/missing-people-stats/docs">Ver documentación</a>
-    </div>
-
-    <div class="api-card">
-      <div class="api-title">Ministry of Justice in Zaragoza</div>
-      <p class="description">
-        Información sobre los delitos tratados en los distintos juzgados de Zaragoza.
-      </p>
-      <a href="/api/v1/ministry-of-justice-in-zaragoza/docs">Ver documentación</a>
-    </div>
-  </div>
-
-  <div class="footer">
-    <p><a href="/about">Ir a la página Sobre el proyecto</a></p>
-  </div>
-</body>
-</html>
