+++ src/frontend/.svelte-kit/adapter-node/nodes/1.js
@@ -3,6 +3,6 @@
 export const index = 1;
 let component_cache;
 export const component = async () => component_cache ??= (await import('../entries/fallbacks/error.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/1.DvqLdR4X.js","_app/immutable/chunks/Bu7QL4ii.js","_app/immutable/chunks/AWa3Cf9u.js","_app/immutable/chunks/DltxqxoW.js","_app/immutable/chunks/BcWd9E2X.js","_app/immutable/chunks/HoNV5Zar.js","_app/immutable/chunks/BXBmFdIA.js","_app/immutable/chunks/ChcIotwt.js","_app/immutable/chunks/D5jAcvew.js","_app/immutable/chunks/C5OmSORT.js"];
+export const imports = ["_app/immutable/nodes/1.DEgpnlxN.js","_app/immutable/chunks/CWj6FrbW.js","_app/immutable/chunks/69_IOA4Y.js","_app/immutable/chunks/DIeogL5L.js","_app/immutable/chunks/CHhmks2n.js","_app/immutable/chunks/Cuedrw2D.js","_app/immutable/chunks/umIrC32l.js","_app/immutable/chunks/Bj6axSiC.js","_app/immutable/chunks/3yENUh-w.js","_app/immutable/chunks/Cwk7wz8z.js","_app/immutable/chunks/BwYDwrZ3.js","_app/immutable/chunks/BuVK3Hzi.js"];
 export const stylesheets = [];
 export const fonts = [];
