+++ src/frontend/.svelte-kit/adapter-node/nodes/4.js
@@ -2,7 +2,7 @@
 
 export const index = 4;
 let component_cache;
-export const component = async () => component_cache ??= (await import('../entries/pages/integrations/graphs-JMRL/g19-accident-rate-stats/_page.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/4.CbnwFcyk.js","_app/immutable/chunks/Bu7QL4ii.js","_app/immutable/chunks/AWa3Cf9u.js","_app/immutable/chunks/DltxqxoW.js","_app/immutable/chunks/HoNV5Zar.js","_app/immutable/chunks/BXBmFdIA.js","_app/immutable/chunks/C5OmSORT.js"];
-export const stylesheets = ["_app/immutable/assets/4.BHNJgykA.css"];
+export const component = async () => component_cache ??= (await import('../entries/pages/integrations/graphs-JMRL/ext1/_page.svelte.js')).default;
+export const imports = ["_app/immutable/nodes/4.CWG1ehzT.js","_app/immutable/chunks/CWj6FrbW.js","_app/immutable/chunks/69_IOA4Y.js","_app/immutable/chunks/DIeogL5L.js"];
+export const stylesheets = [];
 export const fonts = [];
