+++ src/frontend/.svelte-kit/output/client/_app/version.json
@@ -1,1 +1,1 @@
-{"version":"1746727407523"}
+{"version":"1746816398773"}
