+++ src/frontend/.svelte-kit/output/server/.vite/manifest.json
@@ -100,6 +100,30 @@
       "_app/immutable/assets/_page.DrGWq2lj.css"
     ]
   },
+  "src/routes/integrations/graphs-JMRL/ext1/+page.svelte": {
+    "file": "entries/pages/integrations/graphs-JMRL/ext1/_page.svelte.js",
+    "name": "entries/pages/integrations/graphs-JMRL/ext1/_page.svelte",
+    "src": "src/routes/integrations/graphs-JMRL/ext1/+page.svelte",
+    "isEntry": true
+  },
+  "src/routes/integrations/graphs-JMRL/ext2/+page.svelte": {
+    "file": "entries/pages/integrations/graphs-JMRL/ext2/_page.svelte.js",
+    "name": "entries/pages/integrations/graphs-JMRL/ext2/_page.svelte",
+    "src": "src/routes/integrations/graphs-JMRL/ext2/+page.svelte",
+    "isEntry": true
+  },
+  "src/routes/integrations/graphs-JMRL/ext3/+page.svelte": {
+    "file": "entries/pages/integrations/graphs-JMRL/ext3/_page.svelte.js",
+    "name": "entries/pages/integrations/graphs-JMRL/ext3/_page.svelte",
+    "src": "src/routes/integrations/graphs-JMRL/ext3/+page.svelte",
+    "isEntry": true
+  },
+  "src/routes/integrations/graphs-JMRL/g16-unemployment-stats/+page.svelte": {
+    "file": "entries/pages/integrations/graphs-JMRL/g16-unemployment-stats/_page.svelte.js",
+    "name": "entries/pages/integrations/graphs-JMRL/g16-unemployment-stats/_page.svelte",
+    "src": "src/routes/integrations/graphs-JMRL/g16-unemployment-stats/+page.svelte",
+    "isEntry": true
+  },
   "src/routes/integrations/graphs-JMRL/g19-accident-rate-stats/+page.svelte": {
     "file": "entries/pages/integrations/graphs-JMRL/g19-accident-rate-stats/_page.svelte.js",
     "name": "entries/pages/integrations/graphs-JMRL/g19-accident-rate-stats/_page.svelte",
