+++ src/frontend/.svelte-kit/output/server/chunks/internal.js
@@ -548,7 +548,7 @@ const options = {
 		<div class="error">
 			<span class="status">` + status + '</span>\n			<div class="message">\n				<h1>' + message + "</h1>\n			</div>\n		</div>\n	</body>\n</html>\n"
   },
-  version_hash: "gu3fmc"
+  version_hash: "j2olnj"
 };
 async function get_hooks() {
   let handle;
