+++ src/frontend/.svelte-kit/output/server/manifest-full.js
@@ -10,7 +10,7 @@ return {
 	assets: new Set(["favicon.png"]),
 	mimeTypes: {".png":"image/png"},
 	_: {
-		client: {start:"_app/immutable/entry/start.DOZqx2Uq.js",app:"_app/immutable/entry/app.JXA0pYxn.js",imports:["_app/immutable/entry/start.DOZqx2Uq.js","_app/immutable/chunks/ChcIotwt.js","_app/immutable/chunks/AWa3Cf9u.js","_app/immutable/chunks/D5jAcvew.js","_app/immutable/chunks/C5OmSORT.js","_app/immutable/entry/app.JXA0pYxn.js","_app/immutable/chunks/AWa3Cf9u.js","_app/immutable/chunks/BcWd9E2X.js","_app/immutable/chunks/HoNV5Zar.js","_app/immutable/chunks/Bu7QL4ii.js","_app/immutable/chunks/NmNIvlnM.js","_app/immutable/chunks/D5jAcvew.js","_app/immutable/chunks/C5OmSORT.js"],stylesheets:[],fonts:[],uses_env_dynamic_public:false},
+		client: {start:"_app/immutable/entry/start.fMY1n3UV.js",app:"_app/immutable/entry/app.4A_jVm6F.js",imports:["_app/immutable/entry/start.fMY1n3UV.js","_app/immutable/chunks/Cwk7wz8z.js","_app/immutable/chunks/CHhmks2n.js","_app/immutable/chunks/DIeogL5L.js","_app/immutable/chunks/BwYDwrZ3.js","_app/immutable/chunks/BuVK3Hzi.js","_app/immutable/entry/app.4A_jVm6F.js","_app/immutable/chunks/CHhmks2n.js","_app/immutable/chunks/DIeogL5L.js","_app/immutable/chunks/Cuedrw2D.js","_app/immutable/chunks/umIrC32l.js","_app/immutable/chunks/Bj6axSiC.js","_app/immutable/chunks/CWj6FrbW.js","_app/immutable/chunks/LsQLFvbT.js","_app/immutable/chunks/BwYDwrZ3.js","_app/immutable/chunks/BuVK3Hzi.js"],stylesheets:[],fonts:[],uses_env_dynamic_public:false},
 		nodes: [
 			__memo(() => import('./nodes/0.js')),
 			__memo(() => import('./nodes/1.js')),
@@ -18,7 +18,11 @@ return {
 			__memo(() => import('./nodes/3.js')),
 			__memo(() => import('./nodes/4.js')),
 			__memo(() => import('./nodes/5.js')),
-			__memo(() => import('./nodes/6.js'))
+			__memo(() => import('./nodes/6.js')),
+			__memo(() => import('./nodes/7.js')),
+			__memo(() => import('./nodes/8.js')),
+			__memo(() => import('./nodes/9.js')),
+			__memo(() => import('./nodes/10.js'))
 		],
 		routes: [
 			{
@@ -36,24 +40,52 @@ return {
 				endpoint: null
 			},
 			{
+				id: "/integrations/graphs-JMRL/ext1",
+				pattern: /^\/integrations\/graphs-JMRL\/ext1\/?$/,
+				params: [],
+				page: { layouts: [0,], errors: [1,], leaf: 4 },
+				endpoint: null
+			},
+			{
+				id: "/integrations/graphs-JMRL/ext2",
+				pattern: /^\/integrations\/graphs-JMRL\/ext2\/?$/,
+				params: [],
+				page: { layouts: [0,], errors: [1,], leaf: 5 },
+				endpoint: null
+			},
+			{
+				id: "/integrations/graphs-JMRL/ext3",
+				pattern: /^\/integrations\/graphs-JMRL\/ext3\/?$/,
+				params: [],
+				page: { layouts: [0,], errors: [1,], leaf: 6 },
+				endpoint: null
+			},
+			{
+				id: "/integrations/graphs-JMRL/g16-unemployment-stats",
+				pattern: /^\/integrations\/graphs-JMRL\/g16-unemployment-stats\/?$/,
+				params: [],
+				page: { layouts: [0,], errors: [1,], leaf: 7 },
+				endpoint: null
+			},
+			{
 				id: "/integrations/graphs-JMRL/g19-accident-rate-stats",
 				pattern: /^\/integrations\/graphs-JMRL\/g19-accident-rate-stats\/?$/,
 				params: [],
-				page: { layouts: [0,], errors: [1,], leaf: 4 },
+				page: { layouts: [0,], errors: [1,], leaf: 8 },
 				endpoint: null
 			},
 			{
 				id: "/ministry-of-justice-in-zaragoza",
 				pattern: /^\/ministry-of-justice-in-zaragoza\/?$/,
 				params: [],
-				page: { layouts: [0,], errors: [1,], leaf: 5 },
+				page: { layouts: [0,], errors: [1,], leaf: 9 },
 				endpoint: null
 			},
 			{
 				id: "/ministry-of-justice-in-zaragoza/[province]/[year]/[id]",
 				pattern: /^\/ministry-of-justice-in-zaragoza\/([^/]+?)\/([^/]+?)\/([^/]+?)\/?$/,
 				params: [{"name":"province","optional":false,"rest":false,"chained":false},{"name":"year","optional":false,"rest":false,"chained":false},{"name":"id","optional":false,"rest":false,"chained":false}],
-				page: { layouts: [0,], errors: [1,], leaf: 6 },
+				page: { layouts: [0,], errors: [1,], leaf: 10 },
 				endpoint: null
 			}
 		],
