+++ src/frontend/.svelte-kit/output/server/nodes/2.js
@@ -3,6 +3,6 @@
 export const index = 2;
 let component_cache;
 export const component = async () => component_cache ??= (await import('../entries/pages/_page.svelte.js')).default;
-export const imports = ["_app/immutable/nodes/2.BLZjqBgs.js","_app/immutable/chunks/Bu7QL4ii.js","_app/immutable/chunks/AWa3Cf9u.js","_app/immutable/chunks/DltxqxoW.js","_app/immutable/chunks/HoNV5Zar.js"];
+export const imports = ["_app/immutable/nodes/2.DUikPCyO.js","_app/immutable/chunks/CWj6FrbW.js","_app/immutable/chunks/69_IOA4Y.js","_app/immutable/chunks/DIeogL5L.js","_app/immutable/chunks/CHhmks2n.js","_app/immutable/chunks/umIrC32l.js","_app/immutable/chunks/Bj6axSiC.js"];
 export const stylesheets = [];
 export const fonts = [];
